+++ tests/test_preprocessings/test_preprocessings.py
@@ -9,6 +9,7 @@ from binarizer import binarizer
 from function_transformer import function_transformer
 from kernel_centerer import kernel_centerer
 from multilabel_binarizer import multilabel_binarizer
+from max_abs_scaler import max_abs_scaler
 
 PREPROCESSINGS = [one_hot_encoder,
                   label_binarizer,
@@ -18,6 +19,7 @@ PREPROCESSINGS = [one_hot_encoder,
                   function_transformer,
                   kernel_centerer,
                   multilabel_binarizer,
+                  max_abs_scaler,
                   ]
 
 @pytest.fixture(scope="session", autouse=True)